적을 지정하고 적 방향으로 화살 이동

+++ a2dg/src/main/java/kr/ac/tukorea/ge/spgp2026/a2dg/objects/AnimSprite.kt
@@ -60,8 +60,8 @@ open class AnimSprite(
             canvas.scale(
                 -1f,
                 1f,
-                dstRect.centerX().toFloat(),
-                dstRect.centerY().toFloat()
+                dstRect.centerX(),
+                dstRect.centerY()
             )
 
             canvas.drawBitmap(bitmap, srcRect, dstRect, null)
